pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flagevalmm"
version = "0.3.5"
description = "FlagEval multimodal evaluation toolkit"
readme = "README.md"
authors = [
    { name="FlagEval", email="flageval@baai.ac.cn" }
]
license = { text = "MIT" }
dependencies = [
    "numpy>=1.22.2",
    "mmengine>=0.8.0",
    "flask",
    "pillow",
    "transformers",
    "pycocotools",
    "torch>=2.0.0",
    "torch-fidelity>=0.3.0",
    "torchmetrics",
    "opencv_python>=4.8.0.74",
    "easydict",
    "pandas",
    "datasets",
    "tenacity",
    "openai",
    "av",
    "accelerate",
    "word2number"]  
requires-python = ">=3.10"


[project.optional-dependencies]
dev = [
    "pre-commit"
]


[tool.setuptools.packages.find]
include = ["flagevalmm*"]

[project.scripts]
flagevalmm = "flagevalmm.eval:run"
